+++ packages/components/src/Combobox/components/ComboboxTrigger/ComboboxTriggerChip/ComboboxTriggerChip.tsx
@@ -1,9 +1,9 @@
 import React from "react";
+import { ComboboxTriggerChipProps } from "@jobber/components/Combobox/Combobox.types";
 import styles from "./ComboboxTriggerChip.css";
 import { Icon } from "../../../../Icon";
 import { ComboboxContext } from "../../../ComboboxProvider";
 import { Typography } from "../../../../Typography";
-import { ComboboxTriggerChipProps } from "../../../Combobox.types";
 
 export function ComboboxTriggerChip(
   props: ComboboxTriggerChipProps,